chat-django/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coverage==6.2
cryptography==36.0.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.0
django-debug-toolbar==3.2.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-spectacular==0.26.2
duplicity==0.8.21
exceptiongroup==1.1.1
Faker==10.0.0
fasteners==0.14.1
flake8==4.0.1
frozenlist==1.3.3
future==0.18.2
google==3.0.0
gunicorn==20.1.0
gyp==0.1
h11==0.12.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
jeepney==0.7.1
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.45.1
mypy==0.910
mypy-extensions==0.4.3
natsort==8.0.2
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
openai==0.27.2
outcome==1.1.0
packaging==21.3
pandas==1.5.3
parameterized==0.8.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.12.4
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pyflakes==2.4.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyPDF2==3.0.1
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==6.2.5
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pytz==2021.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
schedule==1.2.0
SecretStorage==3.3.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.2
systemd-python==234
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
usb-creator==0.3.7
wadllib==1.3.6
watchdog==2.1.6
websockets==9.1
whitenoise==5.3.0
wsproto==1.0.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarl==1.8.2
yt-dlp==2022.4.8
zipp==1.0.0
